Calculate Coordination Numbers

Counts the number of nearest neighbors for each central atom based on a table of bonded pairs.

Usage

calculate_neighbor_counts(bonded_pairs_table)

Arguments

bonded_pairs_table

A data.table of bonded pairs.

Value

A data.table with columns 'Atom' and 'CoordinationNumber'.

See also